:sparkles: Let jobseekers unlock their accounts
- Uses the Devise module called Lockable.
- Add 'too many attempts' page. We redirect to this page with a before_action
hook if the jobseeker's account is locked, replacing the inbuilt Devise
functionality (which is to reload the sign in page with a flash
notification.)
- Adapt the styling of Devise's unlocks/new view to match our site. This is
the page which Devise redirects to if an unlock token is invalid. I have done
the minimal adaptations to make this page presentable: e.g. not using Times
New Roman. This page is awaiting some content design which can be inserted
into the translations file without breaking things.